Demo almost done. Environment simulation runs successfully and sensor read correct values.
A diagram designating application components and how
they cooperate, constructed.

TODO: Support arbitrary arguments for event handlers

+++ Sense/src/Rabbits.h
@@ -8,9 +8,9 @@
 		} sys;
 		
 		struct _model_rabits {
-			uint8_t lastgen;
-			uint8_t thisgen;
-			uint8_t tmp;
+			uint16_t lastgen;
+			uint16_t thisgen;
+			uint16_t tmp;
 			
 		} rabbits;
 	};
